+++ internal/entities/localfile_helper_test.go
@@ -7,12 +7,12 @@ import (
 	"testing"
 )
 
-var DirPath = "E:\\Anime"
-var LocalFilePath = "E:\\Anime\\Bungou Stray Dogs 5th Season\\[SubsPlease] Bungou Stray Dogs - 61 (1080p) [F609B947].mkv"
+var dirPath = "E:\\Anime"
+var testFilePath = "E:\\Anime\\Bungou Stray Dogs 5th Season\\[SubsPlease] Bungou Stray Dogs - 61 (1080p) [F609B947].mkv"
 
 func TestLocalFile_GetTitleVariations(t *testing.T) {
 
-	lf := NewLocalFile(LocalFilePath, DirPath)
+	lf := NewLocalFile(testFilePath, dirPath)
 
 	if assert.NotNil(t, lf) {
 		tv := lo.Map(lf.GetTitleVariations(), func(item *string, _ int) string { return *item })
@@ -22,6 +22,20 @@ func TestLocalFile_GetTitleVariations(t *testing.T) {
 		}) {
 			t.Log(spew.Sdump(lf.GetTitleVariations()))
 		}
+	}
+
+}
+
+var testFilePath2 = "E:\\Anime\\Shakugan No Shana\\Shakugan No Shana I\\Opening\\OP01.mkv"
+
+func TestLocalFile_GetTitleVariations2(t *testing.T) {
+
+	lf := NewLocalFile(testFilePath2, dirPath)
+
+	if assert.NotNil(t, lf) {
+		tv := lo.Map(lf.GetTitleVariations(), func(item *string, _ int) string { return *item })
+
+		spew.Dump(tv)
 	}
 
 }