+++ seanime-web/src/lib/server/types.ts
@@ -61,6 +61,7 @@ export type MediaPlayerSettings = {
     vlcPath: string
     mpcPort: number
     mpcPath: string
+    mpvSocket: string
 }
 
 export type LibrarySettings = {